+++ test/ReactPlayer/props.tsx
@@ -1,16 +1,18 @@
+import '../helpers/server-safe-globals.js';
 import { test } from 'zora';
 import React from 'react';
 import { create } from 'react-test-renderer';
 import ReactPlayer from '../../src/index';
+import { render } from '../helpers/helpers';
 
-test('className', (t) => {
-  const wrapper = create(<ReactPlayer className="react-player" />);
-  t.equal(wrapper.root.findByType('div').props.className, 'react-player');
+test('className', async (t) => {
+  const wrapper = render(<ReactPlayer src="file.mp4" className="react-player" />);
+  t.equal(wrapper.root.findByType('video').props.className, 'react-player');
 });
 
 test('style', (t) => {
-  const wrapper = create(<ReactPlayer style={{ marginTop: '1rem' }} />);
-  t.equal(wrapper.root.findByType('div').props.style.marginTop, '1rem');
+  const wrapper = render(<ReactPlayer src="file.mp4" style={{ marginTop: '1rem' }} />);
+  t.equal(wrapper.root.findByType('video').props.style.marginTop, '1rem');
 });
 
 test('wrapper - string', (t) => {
